requirements.txt
beautifulsoup4==4.5.1
defusedxml==0.4.1
Django==1.10.4
django-allauth==0.29.0
django-annoying==0.10.3
django-bootstrap-pagination==1.6.2
django-braces==1.10.0
django-crispy-forms==1.6.1
django-environ==0.4.1
django-model-utils==2.6
django-scheduler==0.8.2
django-taggit==0.21.3
icalendar==3.11
konlpy==0.4.4
oauthlib==2.0.1
psycopg2==2.6.2
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
